refactored package for scheduler
