+++ build.gradle
@@ -25,7 +25,7 @@ subprojects {
 
   [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
-  mainClassName = "com.apache.mesos.hdfs.Main"
+  mainClassName = "org.apache.mesos.hdfs.scheduler.Main"
 
   group = "com.apache.mesos.hdfs"
   version = "0.1.1"
